requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
Faker==33.0.0
pytest==8.3.3
requests==2.32.3
